Made new controllers package and moved all controllers there (to break up views package).
