+++ src/main/java/controllers/LoginController.java
@@ -1,4 +1,4 @@
-package views;
+package controllers;
 
 
 import user_login_use_case.LoginData;
