+++ src/main/java/controllers/MessageDeleteController.java
@@ -1,4 +1,4 @@
-package views;
+package controllers;
 
 
 import message_edit_delete_use_case.*;
