+++ src/main/java/controllers/MessageEditController.java
@@ -1,4 +1,4 @@
-package views;
+package controllers;
 
 import message_edit_delete_use_case.MessageEditData;
 import message_edit_delete_use_case.MessageEditInputBoundary;
