+++ src/main/java/controllers/MessageSearchController.java
@@ -1,4 +1,4 @@
-package views;
+package controllers;
 
 import message_search_use_case.MessageSearchData;
 import message_search_use_case.MessageSearchInputBoundary;
@@ -15,7 +15,7 @@ public class MessageSearchController {
         this.messageSearchInput = messageSearchInput;
     }
 
-    MessageSearchResponse search(MessageSearchData data) {
+    public MessageSearchResponse search(MessageSearchData data) {
         return this.messageSearchInput.search(data);
     }
 }
