+++ src/main/java/views/ChatScreen.java
@@ -1,5 +1,6 @@
 package views;
 
+import controllers.*;
 import entities.Message;
