+++ src/main/java/views/ProfileScreen.java
@@ -1,5 +1,6 @@
 package views;
 
+import controllers.CustomizationController;
 import entities.User;
