+++ src/main/java/views/RecordButton.java
@@ -1,10 +1,12 @@
 package views;
 
 import audio_converter_use_case.AudioConvertData;
+import controllers.AudioConvertController;
+import controllers.AudioRecorderController;
+import controllers.MessageTranslateController;
 import translate_use_case.MessageTranslateData;
 
 import javax.swing.*;
-import javax.swing.border.Border;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -26,13 +28,9 @@ public class RecordButton extends JPanel implements ActionListener {
     static MessageTranslateController MTC;
     boolean recording;
 
-
-
-    ImageIcon micImage;
-
     public RecordButton(AudioRecorderController ARC, AudioConvertController ACC, MessageTranslateController MTC, String lang, JTextField textBox){
 
-        this.MTC = MTC; //#TODO remove after
+        RecordButton.MTC = MTC; //#TODO remove after
         this.lang = lang;
         this.setLayout(new FlowLayout())    ;
         this.recording = false;
@@ -65,36 +63,26 @@ public void display() throws IOException {
         if (recording) {
 
             System.out.println("hihihihi");
-            Thread counter = new Thread(new Runnable() {
-                public void run() {
-                    try {
-                        for (int i = 1; i <= 15-1; i++) {
-                            if (!recording) {
-                                break;
-                            }
-                            Thread.sleep(1000);//The thread sleeps for as long as we record, this determines how long our recording is
+            Thread counter = new Thread(() -> {
+                try {
+                    for (int i = 1; i <= 15-1; i++) {
+                        if (!recording) {
+                            break;
                         }
-                        System.out.println("yyayayay");
-                    } catch (InterruptedException ex) {
-                        ex.printStackTrace();
+                        Thread.sleep(1000);//The thread sleeps for as long as we record, this determines how long our recording is
                     }
-                    try {
-                        display();
-                    } catch (IOException ex) {
-                        throw new RuntimeException(ex);
-                    }
-                    recording = false;
+                    System.out.println("yyayayay");
+                } catch (InterruptedException ex) {
+                    ex.printStackTrace();
                 }
-
-
-            });
-            Thread record = new Thread(new Runnable() {
-                public void run() {audioRecorderController.record();
-
+                try {
+                    display();
+                } catch (IOException ex) {
+                    throw new RuntimeException(ex);
                 }
-
-
+                recording = false;
             });
+            Thread record = new Thread(() -> audioRecorderController.record());
             counter.start();
 
             record.start();
