+++ src/test/java/views/MessageDeleteControllerTest.java
@@ -3,6 +3,7 @@ package views;
 import com.google.cloud.firestore.DocumentReference;
 import com.google.cloud.firestore.Firestore;
 import com.google.firebase.cloud.FirestoreClient;
+import controllers.MessageDeleteController;
 import gateways.MessageDeleteFirebaseSystem;
 import message_edit_delete_use_case.*;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,6 +12,7 @@ import services.DBInitializer;
 
 import java.io.FileNotFoundException;
 import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.ExecutionException;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -34,7 +36,7 @@ class MessageDeleteControllerTest {
         Firestore dbFirestore = FirestoreClient.getFirestore();
         controller.delete(new MessageDeleteData(5, 4));
         DocumentReference chatref = dbFirestore.collection("chats").document("id"+4);
-        assertEquals(0, ( (List<DocumentReference>)chatref.get().get().getData().get("messages")).size());
+        assertEquals(0, ( (List<DocumentReference>) Objects.requireNonNull(chatref.get().get().getData()).get("messages")).size());
 
     }
 }
