+++ src/test/java/views/MessageSearchControllerTest.java
@@ -1,5 +1,6 @@
 package views;
 
+import controllers.MessageSearchController;
 import gateways.MessageSearchFirebaseSystem;
 import message_search_use_case.*;
 import org.junit.jupiter.api.BeforeEach;
